server/requirements.txt
flask==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.31.0
flask-socketio==5.3.6
flask-cors==4.0.0
